refactor(settings): wire new dialog and remove superseded containers
