+++ client/src/components/Nav/SettingsTabs/index.ts
@@ -1,9 +1,2 @@
-export { default as Chat } from './Chat/Chat';
-export { default as Data } from './Data/Data';
-export { default as Speech } from './Speech/Speech';
-export { default as Balance } from './Balance/Balance';
-export { default as General } from './General/General';
-export { default as Account } from './Account/Account';
-export { default as Commands } from './Commands/Commands';
-export { default as Personalization } from './Personalization';
 export { default as About } from './About/About';
+export { ThemeSelector, LangSelector } from './General/Selectors';
